+++ bin/usb.c
@@ -5,8 +5,6 @@
 #include <string.h>
 #include "debug/printf.h"
 
-void delay(uint32_t);
-
 //#define LOG_SIZE  20
 //uint32_t transfer_log_head=0;
 //uint32_t transfer_log_count=0;
@@ -97,6 +95,9 @@ static void endpoint0_complete(void);
 
 static void run_callbacks(endpoint_t *ep);
 
+int32_t usb_device_is_configured(void) {
+	return !!usb_configuration;
+}
 
 FLASHMEM void usb_init(void)
 {
@@ -143,14 +144,12 @@ FLASHMEM void usb_init(void)
 		USBPHY1_CTRL_CLR = USBPHY_CTRL_SFTRST; // reset PHY
 		//USB1_USBSTS = USB1_USBSTS; // TODO: is this needed?
 		printf("USB reset took %d loops\n", count);
-		//delay(10);
 		//printf("\n");
 		//printf("USBPHY1_PWD=%08lX\n", USBPHY1_PWD);
 		//printf("USBPHY1_TX=%08lX\n", USBPHY1_TX);
 		//printf("USBPHY1_RX=%08lX\n", USBPHY1_RX);
 		//printf("USBPHY1_CTRL=%08lX\n", USBPHY1_CTRL);
 		//printf("USB1_USBMODE=%08lX\n", USB1_USBMODE);
-		delay(25);
 	}
 #endif
 	// Device Controller Initialization, page 3161
@@ -197,8 +196,20 @@ FLASHMEM void usb_init(void)
 	//transfer_log_count = 0;
 }
 
+#define POLL_RX_ENDPOINT_MASK(ep) (1 << ep)
+#define POLL_TX_ENDPOINT_MASK(ep) (1 << (ep + 16))
 
-void isr(void)
+//
+// Keep these in sync with the constants
+// in bindings.rs.
+//
+
+typedef enum {
+	POLL_CDC_RX_COMPLETE = 1,
+	POLL_CDC_TX_COMPLETE = 2,
+} poll_flag_t;
+
+uint32_t poll(void)
 {
 	//printf("*");
 
@@ -206,6 +217,7 @@ void isr(void)
 	//  status port reset, suspend, and current connect status.
 	uint32_t status = USB1_USBSTS;
 	USB1_USBSTS = status;
+	uint32_t poll_flags = 0;
 
 	// USB_USBSTS_SLI - set to 1 when enters a suspend state from an active state
 	// USB_USBSTS_SRI - set at start of frame
@@ -232,6 +244,10 @@ void isr(void)
 			setupstatus = USB1_ENDPTSETUPSTAT; // page 3175
 		}
 		uint32_t completestatus = USB1_ENDPTCOMPLETE;
+
+		poll_flags |= (POLL_RX_ENDPOINT_MASK(CDC_RX_ENDPOINT) & completestatus) ? POLL_CDC_RX_COMPLETE : 0;
+		poll_flags |= (POLL_TX_ENDPOINT_MASK(CDC_TX_ENDPOINT) & completestatus) ? POLL_CDC_TX_COMPLETE : 0;
+
 		if (completestatus) {
 			USB1_ENDPTCOMPLETE = completestatus;
 			//printf("USB1_ENDPTCOMPLETE=%lX\n", completestatus);
@@ -308,6 +324,8 @@ void isr(void)
 			USB1_USBINTR &= ~USB_USBINTR_SRE;
 		}
 	}
+
+	return poll_flags;
 }
 
 
@@ -450,7 +468,6 @@ static void endpoint0_setup(uint64_t setupdata)
 		break;
 #if defined(CDC_STATUS_INTERFACE)
 	  case 0x2221: // CDC_SET_CONTROL_LINE_STATE
-		usb_cdc_line_rtsdtr_millis = systick_millis_count;
 		usb_cdc_line_rtsdtr = setup.wValue;
 		__attribute__((fallthrough));
 	  case 0x2321: // CDC_SEND_BREAK