+++ bin/usb_serial.c
@@ -42,7 +42,6 @@ void yield(void);
 //#if F_CPU >= 20000000
 
 uint32_t usb_cdc_line_coding[2];
-volatile uint32_t usb_cdc_line_rtsdtr_millis;
 volatile uint8_t usb_cdc_line_rtsdtr=0;
 volatile uint8_t usb_cdc_transmit_flush_timer=0;
 
@@ -99,7 +98,7 @@ void usb_serial_configure(void)
 	}
 	memset(tx_transfer, 0, sizeof(tx_transfer));
 	tx_head = 0;
-	tx_available = 0;
+	tx_available = TX_SIZE;
 	memset(rx_transfer, 0, sizeof(rx_transfer));
 	memset(rx_count, 0, sizeof(rx_count));
 	memset(rx_index, 0, sizeof(rx_index));
@@ -267,20 +266,11 @@ int usb_serial_getchar(void)
 #define TX_TIMEOUT_MSEC 120
 
 
-// When we've suffered the transmit timeout, don't wait again until the computer
-// begins accepting data.  If no software is running to receive, we'll just discard
-// data as rapidly as Serial.print() can generate it, until there's something to
-// actually receive it.
-static uint8_t transmit_previous_timeout=0;
-
-
 // transmit a character.  0 returned on success, -1 on error
 int usb_serial_putchar(uint8_t c)
 {
 	return usb_serial_write(&c, 1);
 }
-
-extern volatile uint32_t systick_millis_count;
 
 static void timer_config(void (*callback)(void), uint32_t microseconds);
 static void timer_start_oneshot();
@@ -305,50 +295,20 @@ static void timer_stop(void)
 	USB1_GPTIMER0CTRL = 0;
 }
 
+// Keep this in sync with the error constants
+// in bindings.rs.
+typedef enum {
+	SERIAL_NOT_CONFIGURED = -1,
+} serial_err_t;
 
 int usb_serial_write(const void *buffer, uint32_t size)
 {
 	uint32_t sent=0;
 	const uint8_t *data = (const uint8_t *)buffer;
 
-	if (!usb_configuration) return 0;
+	if (!usb_configuration) return SERIAL_NOT_CONFIGURED;
 	while (size > 0) {
 		transfer_t *xfer = tx_transfer + tx_head;
-		int waiting=0;
-		uint32_t wait_begin_at=0;
-		while (!tx_available) {
-			//digitalWriteFast(3, HIGH);
-			uint32_t status = usb_transfer_status(xfer);
-			if (!(status & 0x80)) {
-				if (status & 0x68) {
-					// TODO: what if status has errors???
-					printf("ERROR status = %x, i=%d, ms=%u\n",
-						status, tx_head, systick_millis_count);
-				}
-				tx_available = TX_SIZE;
-				transmit_previous_timeout = 0;
-				break;
-			}
-			if (!waiting) {
-				wait_begin_at = systick_millis_count;
-				waiting = 1;
-			}
-			if (transmit_previous_timeout) return sent;
-			if (systick_millis_count - wait_begin_at > TX_TIMEOUT_MSEC) {
-				// waited too long, assume the USB host isn't listening
-				transmit_previous_timeout = 1;
-				return sent;
-				//printf("\nstop, waited too long\n");
-				//printf("status = %x\n", status);
-				//printf("tx head=%d\n", tx_head);
-				//printf("TXFILLTUNING=%08lX\n", USB1_TXFILLTUNING);
-				//usb_print_transfer_log();
-				//while (1) ;
-			}
-			if (!usb_configuration) return sent;
-			yield();
-		}
-		//digitalWriteFast(3, LOW);
 		uint8_t *txdata = txbuffer + (tx_head * TX_SIZE) + (TX_SIZE - tx_available);
 		if (size >= tx_available) {
 			memcpy(txdata, data, tx_available);
@@ -362,7 +322,7 @@ int usb_serial_write(const void *buffer, uint32_t size)
 			size -= tx_available;
 			sent += tx_available;
 			data += tx_available;
-			tx_available = 0;
+			tx_available = TX_SIZE;
 			timer_stop();
 		} else {
 			memcpy(txdata, data, size);
@@ -391,7 +351,6 @@ void usb_serial_flush_output(void)
 {
 
 	if (!usb_configuration) return;
-	if (tx_available == 0) return;
 	tx_noautoflush = 1;
 	transfer_t *xfer = tx_transfer + tx_head;
 	uint8_t *txbuf = txbuffer + (tx_head * TX_SIZE);
@@ -400,7 +359,7 @@ void usb_serial_flush_output(void)
 	arm_dcache_flush_delete(txbuf, txnum);
 	usb_transmit(CDC_TX_ENDPOINT, xfer);
 	if (++tx_head >= TX_NUM) tx_head = 0;
-	tx_available = 0;
+	tx_available = TX_SIZE;
 	tx_noautoflush = 0;
 }
 
@@ -408,7 +367,6 @@ static void usb_serial_flush_callback(void)
 {
 	if (tx_noautoflush) return;
 	if (!usb_configuration) return;
-	if (tx_available == 0) return;
 	//printf("flush callback, %d bytes\n", TX_SIZE - tx_available);
 	transfer_t *xfer = tx_transfer + tx_head;
 	uint8_t *txbuf = txbuffer + (tx_head * TX_SIZE);
@@ -417,7 +375,7 @@ static void usb_serial_flush_callback(void)
 	arm_dcache_flush_delete(txbuf, txnum);
 	usb_transmit(CDC_TX_ENDPOINT, xfer);
 	if (++tx_head >= TX_NUM) tx_head = 0;
-	tx_available = 0;
+	tx_available = TX_SIZE;
 }