+++ examples/bsp/src/bin/dma_spi.rs
@@ -26,6 +26,8 @@
 #![no_main]
 #![no_std]
 
+mod usb_io;
+
 use teensy4_panic as _;
 
 use bsp::hal::dma;
@@ -159,7 +161,7 @@ static mut HARDWARE_FLAG: Option<HardwareFlag> = None;
 fn main() -> ! {
     let mut peripherals = bsp::Peripherals::take().unwrap();
     let mut systick = bsp::SysTick::new(cortex_m::Peripherals::take().unwrap().SYST);
-    bsp::usb::init(&systick, Default::default()).unwrap();
+    usb_io::init().unwrap();
     let pins = bsp::t40::into_pins(peripherals.iomuxc);
 
     peripherals.ccm.pll1.set_arm_clock(