+++ examples/bsp/src/bin/i2c.rs
@@ -17,6 +17,8 @@
 #![no_std]
 #![no_main]
 
+mod usb_io;
+
 use teensy4_panic as _;
 
 use bsp::hal::i2c::ClockSpeed;
@@ -45,7 +47,7 @@ fn main() -> ! {
     let mut peripherals = bsp::Peripherals::take().unwrap();
     let mut systick = bsp::SysTick::new(cortex_m::Peripherals::take().unwrap().SYST);
     let pins = bsp::t40::into_pins(peripherals.iomuxc);
-    bsp::usb::init(&systick, Default::default()).unwrap();
+    usb_io::init().unwrap();
     systick.delay(5000);
 
     log::info!("Enabling I2C clocks...");