+++ examples/bsp/src/bin/pwm.rs
@@ -13,6 +13,8 @@
 #![no_std]
 #![no_main]
 
+mod usb_io;
+
 use teensy4_panic as _;
 
 use bsp::hal::pwm::Channel;
@@ -31,8 +33,7 @@ fn main() -> ! {
     let mut p = bsp::Peripherals::take().unwrap();
     let mut systick = bsp::SysTick::new(cortex_m::Peripherals::take().unwrap().SYST);
     let pins = bsp::t40::into_pins(p.iomuxc);
-    // Initialize the logging, so we can use it in the PWM loop below
-    bsp::usb::init(&systick, Default::default()).unwrap();
+    usb_io::init().unwrap();
     // Delay is only to let a user set-up their USB serial connection...
     systick.delay(5000);
     // Set the core and IPG clock. The IPG clock frequency drives the PWM (sub)modules