+++ examples/bsp/src/bin/spi.rs
@@ -21,12 +21,13 @@
 #![no_std]
 #![no_main]
 
+mod usb_io;
+
 use teensy4_panic as _;
 
 use cortex_m_rt::entry;
-use teensy4_bsp as bsp;
-
 use embedded_hal::{blocking::spi::Transfer, digital::v2::OutputPin};
+use teensy4_bsp as bsp;
 
 const SPI_BAUD_RATE_HZ: u32 = 1_000_000;
 
@@ -34,7 +35,7 @@ const SPI_BAUD_RATE_HZ: u32 = 1_000_000;
 fn main() -> ! {
     let mut peripherals = bsp::Peripherals::take().unwrap();
     let mut systick = bsp::SysTick::new(cortex_m::Peripherals::take().unwrap().SYST);
-    bsp::usb::init(&systick, Default::default()).unwrap();
+    usb_io::init().unwrap();
     let pins = bsp::t40::into_pins(peripherals.iomuxc);
 
     peripherals.ccm.pll1.set_arm_clock(